Refactor viz component
